[OT-233][FIX]: 북마크 부분 타입 위치 변경
