+++ src/entities/bookmark/api/bookmarkApi.ts
@@ -1,5 +1,36 @@
 import { api } from "@shared/api";
-import { ApiResponse, BookmarkContentResponse, BookmarkShortFormResponse } from "@shared/types";
+import { ApiResponse, PageInfo } from "@shared/types";
+
+// 북마크 콘텐츠 dataList 안쪽 타입
+export interface BookmarkContentItem {
+  mediaId: number;
+  mediaType: "CONTENTS" | "SERIES";
+  title: string;
+  description: string;
+  posterUrl: string;
+  positionSec: number;
+  duration: number;
+}
+
+// 북마크 숏폼 dataList 안쪽 타입
+export interface BookmarkShortFormItem {
+  mediaId: number;
+  title: string;
+  description: string;
+  thumbnailUrl: string;
+}
+
+// 북마크 콘텐츠 조회 목록 전체 틀
+export interface BookmarkContentResponse {
+  pageInfo: PageInfo;
+  dataList: BookmarkContentItem[];
+}
+
+// 북마크 숏폼 조회 목록 전체 틀
+export interface BookmarkShortFormResponse {
+  pageInfo: PageInfo;
+  dataList: BookmarkShortFormItem[];
+}
 
 export const bookmarkApi = {
   getBookmarkContents: async (page: number) =>
